feat(cli): Add completion generator
